requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic_settings
psycopg2
python-multipart
pillow